Modified sampling algorithm

+++ sampling.hpp
@@ -4,6 +4,6 @@
 #include <stdint.h>
 #include "HX711.h"
 
-int16_t getweight(HX711, uint16_t);
+float getweight(HX711, uint16_t);
 
 #endif
